egressgwha test support multiple egress groups
This commit introduces support for multiple egress groups in a
egressgwha policy. Prior to this, we define a single egress group within
each policy. To align with real-world policies with multiple groups as in
`v1.IsovalentEgressGatewayPolicySpec`, this commit moves the needed data
into a `egressGroupParams` struct, and refactors the related test functions
to accept a group ID (index in the list).

Fixes: #4481

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>