egressgwha add test gw in multiple egress groups
This commit introduces a test for checking the behavior of the manager in
case a node being selected as active gateway for multiple egress groups
within the same policy. The current behavior after #5056 is to log an
error message and stop the regeneration of the gateway config for the
current group, preserving the previous valid value. For instance, when a
node is selected as active gateway for both egress group1 (egressIP1) and
group2 (egressIP2), the gateway config generation is successful for group1
(egressIP1), but it fails for the group2 (egressIP2) with the log message,
ensuring that only the previous value is used/advertised/in-map (egressIP1).

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>