lb: add t1-only backend compatibility
Currently, if a T1 only TCP application references a backend with
different ports on the backend addresses, this results in an endless
reconciliation loop trying to create the k8s Endpoint.

Even though exponential backoff is in place, the user that created the
resources doesn't have a feedback.

Therefore, this commit adds validation for T1 only backend validation
to the LBService condition that is already validating the backend
compatibility.

The reconciliation no longer throws an error - instead the address gets
skipped and a debug log is emited.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>