lb: validate max 1 tcp proxy route
Currently, the implementation of TCPProxy application only supports
one route. The number of routes for a TCP Proxy application on
an LBService isn't validated except that it needs to be one.

If there were multiple of them, the logic just skippes the route.

This commit changes this by validating that there's only one route
set per TCP application via CRD validation. This way we can get rid
of the internal logic.

Once we support multiple TCPProxy routes (e.g. by source IP), we can change
this.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>