enterprise/encryption-policy: Fix clustermesh support
This commit fixes an issue where the encryption policy engine did not
receive identities allocated in remote clusters. The fix subscribes to
the new global identity handler (added in a previous commit). To avoid
circular dependencies between the network policy trifecta and the
encryption policy engine, and in order to keep using the `job.Observer`
logic for identity updates, the values received by the callback are
buffered and made available via a custom observable.

In addition, this also adds a new initializer for the clustermesh sync
event. The previous identity sync initializer for local identities is
called after we have observed the first batch of identity changes.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>