lb: json format for file accesslogger
Currently, the main reason for writing the T2 Envoy accesslog to
a file is that it can be picked up by a logforwarder (e.g. fluentd)
and sent to a logging backend. Most of these tools prefer structured
JSON as the logging format.

Therefore, this commit introduces structured JSON logging format for the T2
Envoy access logger that writes to the file.

The accesslogger that writes to stdout (Envoy Pod log) still uses the
non-JSON format.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>