fqdn-proxy: remove active connection mgmt
This removes all logic which tries to actively manage the gRPC
connection as if it was a net.Conn. A gRPC client conn does not know
unrecoverable errors! It is only "dead" when in Shutdown state, which it
only reaches _by choice of the application_. Said differently, _only_ if
the application calls conn.Close(), all future RPCs will fail without
attempting to reach the remote.

In addition, the (experimental) API used to track/manage the connection
state is not very usable. WaitForStateChange is unfortunate in that it
does not wait until the target state is _reached_, but rather until the
target state is not/no longer the current state. It's hence trivial to
miss transitions, since the states cycle quickly.

The original intent of the active connection management was likely to
learn that connectivity to the agent broke quickly. To do so, however,
it's enough to specify more aggressive connection and backoff
parameters, since the defaults are oriented towards gRPC over the
network.

As part of this, clean up the creation of the gRPC connection by
dropping uses of deprecated grpc.Dial/grpc.WithBlock.

One of the reasons Dial is deprecated is that it invokes the wrong layer
of abstraction - a gRPC client connection is _not_ like a net.Conn which
needs to be actively managed.

WithBlock is deprecated because it tries to force gRPC to establish a
connection on client creation. This goes against the principle of the
client to manage its own connection, which attempts establishment on
first RPC. Since a connection can fail at any point, the RPC needs to
handle a connection error anyway so we don't gain anything by trying to
force a connect early.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>