egressgwha: Add support for IPAM and AZ affinity compatibility
The maxGatewayNodes parameter in the IsovalentEgressGatewayPolicy has a
different semantic if AZ affinity is enabled.

When AZ affinity is disabled, maxGatewayNodes is the maximum number of
the active gateways selected by the operator.  Conversely, when AZ
affinity is enabled, maxGatewayNodes is the maximum number of active
gateways for each affinity zone selected by the operator.

Thus, in the latter case, the total number of active gateways from all
affinity zones, listed in activeGatewayIPsByAZ, might be greater then
the number of active gateways listed in activeGatewayIPs. On the agent
side, for a policy relying on AZ affinity, activeGayewayIPsByAZ is
authoritative when reconciling the gateway BPF egress maps, while
activeGatewayIPs is simply disregarded.

Therefore, when IPAM is used in conjunction with AZ affinity, to be sure
to allocate an egress IP to each active gateway, the operator should
consider all the gateways listed in activeGatewayIPsByAZ instead of the
ones in activeGatewayIPs.

Moreover, when the operator cannot fulfill all the egress IP allocation
requests (e.g: the specified egress CIDR does not have enough addresses
to allocate), it should remove the gateways without an allocated IP from
both activeGatewayIPs and activeGatewayIPsByAZ.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>