egressgwha: Sort healthy and active gateway IPs in IEGP Status
Sort all the gateway IPs lists in IEGP Status, specifically
activeGatewayIPs, activeGatewayIPsByAZ and healthyGatewayIPs. This helps
both in terms of glanceability when debugging the operator gateways
selection in real clusters and also in making the unit tests easier to
write, specifically when comparing the resulting gateway lists with the
expected ones.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>