fqdn-proxy: track expected agent connectivity
There's no point in actively trying to manage our gRPC connection, but
it's still useful to know whether the last RPC to the agent succeeded,
so that we can log (once) when connectivity is lost or established.

The gRPC connectivity states are still not the right tool for the job -
instead, look at the errors returned by the RPCs to determine whether we
encountered no error/an error in gRPC/an error at the application layer.

The middle case is the one where the agent connection is broken, since
application errors in the called procedures still indicate that the call
worked.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>