CI: add workflow to run e2e tests against vSphere LoadBalancers
Add "Enterprise LoadBalancer vSphere" workflow to execute the enterprise
LB end-to-end tests against LoadBalancers provisioned via the
isovalent-operator.

This workflow runs on a self-hosted runner with the tags "self-hosted"
and "vsphere". The runner is an Ubuntu 22.04 VM running within the
vSphere lab environment. It is preconfigured with a kubeconfig for a
CI-only K8s cluster running cluster-api-vsphere (capv) and the
isovalent-operator for deploying enterprise LoadBalancers.

See the README in isovalent-operator for more details, including an
architecture diagram:

https://github.com/isovalent/isovalent-operator/tree/pr/docs?tab=readme-ov-file#ci-running-in-vsphere

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>