lb: Implement HTTP remote JWKS fetching
Support fetching remote JWKS with HTTP. In Envoy, it can be achieved
with the configuration like following.

provider0:
  remoteJwks:
    httpUri:
      cluster: example.com
      uri: "http://example.com:8080/oauth/v1/certs"

On the other hand, the user's input for IsovalentLB is like the
following:

jwt:
  jwks:
    httpURI:
      uri: http://example.com:8080/oauth/v1/certs

Currently, we don't let users to specify cluster. Therefore, users
cannot specify Cluster properties such as custom TLS certificate, custom
DNS resolver, or TLS parameters. We can potentially introduce
configuration knobs for the individual parameters or reference to the
LBBackendPool object under httpURI in the future.

For the moment, we generate the bare-minimal Cluster like the following:

name: jwks_cluster_http_provider0
type: STRICT_DNS
loadAssignment:
  clusterName: jwks_cluster_http_provider1
  endpoints:
  - lbEndpoints:
    - endpoint:
        address:
          socketAddress:
            address: example.com
            portValue: 443
transportSocket:
  name: envoy.transport_sockets.tls
  typedConfig:
    '@type': type.googleapis.com/envoy.extensions.transport_sockets.tls.v3.UpstreamTlsContext

The cluster is generated for each application + provider combination. It
automatically determins the plain-text vs TLS based on the URI scheme
and generate transportSocket configuration. The cluster type is always
STRICT_DNS even if the hostname in the URI is an IP address. If the
user specifies the port number, it will be also reflected.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>