chore: Add escalation privileges to the operator service account
To grant privileges to an other service account the granters
need to either have the privileges they want to allocate or to
have the verbs `escalate` and `bind`.
As helm charts evolve it would be tedious to keep the permissions
needed by the operator aligned. Also, as cilium is privileged
the required permissions would already provide escalation
paths, so that allowing privilege escalation through RBAC is
not a significant degration of the securty posture compared
to the alternative.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>