egressgwha: Improve IPAM allocation strategy with AZ affinity
Improve the IPAM allocation algorithm prioritizing affinity zones that
have fewer allocations. This allows to balance the allocatable addresses
among all the affinity zones and, if available addresses are not enough
to satisfy all requests (i.e: not enough for all the active gateways in
all AZs), to cover as much affinity zones as possible.

For each allocation round, the algorithm consider all the zones with the
smallest number of allocations, and assign one egress IP to each of
those, evaluating them in lexicographical order to keep the algorithm
consistent.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>