egressgwha: Rebalance IPAM allocations to cover all AZs
The IPAM allocation algorithm strives to keep previous egress IPs
allocations stable, in order to minimize traffic disruption for already
established connections. To do so, the first step in the allocation
logic is to reserve in advance all those egress IPs from the pool.

In case of a policy where the total number of allocations exceeds the
pool capacity, doing those initial reservations may lead to an
imbalance. Specifically, we may end up with an affinity zone having many
allocated IPs, being assigned from a previous policy version and now
reserved "in advance", while there may be one or more "starving"
affinity zones (i.e: zones without any IPs).

As an example, consider a pool of 4 available addresses. The previous
iteration of the policy led to the following allocations:

| Affinity Zones | Active Gateway IPs |        Egress IP         |
|      ---       |        ---         |           ---            |
|      az_1      | 10.0.0.1, 10.0.0.2 | 192.168.1.1, 192.168.1.3 |
|      az_2      | 10.0.0.3, 10.0.0.4 | 192.168.1.2, 192.168.1.4 |

Now, suppose that the policy is changed so that we have a new affinity
zone "az_3", with two new active gateways. Due to the early
reservations, the new affinity zone is left without any allocation:

| Affinity Zones | Active Gateway IPs |        Egress IP         |
|      ---       |        ---         |           ---            |
|      az_1      | 10.0.0.1, 10.0.0.2 | 192.168.1.1, 192.168.1.3 |
|      az_2      | 10.0.0.3, 10.0.0.4 | 192.168.1.2, 192.168.1.4 |
|      az_3      | 10.0.0.5, 10.0.0.6 |                          |

To improve the affinity zones coverage in this scenario, we first check
if the total number of allocated addresses (equal to the maximum
capacity of the pool) is enough to assign at least one to each affinity
zone. If it isn't, then we leave the allocations unaltered, in order to
avoid connections disruption.  Instead, if there are enough addresses,
then we release one address at a time from the zone with the highest
number of allocations and we assign that to a zone without any IP,
repeating this handoff until all affinity zones have at least one
address allocated.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>