tests: Test ingestion of K8sEvents in Timescape￼
This commit adds an e2e test to check that Hubble Timescape is able to
ingest NetworkPolicy change events through the integrated Hubble k8s
exporter.

To do this we introduce a new `go.mod` in `enterprise/hubble-timescape`.
We do this because, in E2E tests we want to include the timescape
version we actually ship in the helm chart. As we only support a single
timescape version, this will be the latest version, even for Cilium
stable versions. If we'd include the latest timescape as a Go dependency
in `isovalent/cilium`, this will bump arbitrary dependencies in stable
versions and will very likely introduce issues.

By moving the Timescape tests to a separate Go module, Timesacpe test
dependencies will not impact `isovalent/cilium`

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>