lb: Remove -cleanup=false for multi-node test
Currently, the multi-node tests are pretty flaky. Since there's no
staticsics for individual test cases, it's hard to say the "especially
flaky" test. However, in my observation, the failing tests are pretty
ramdom.

```
$ ci-dashboard show isovalent cilium -b main-ce -w enterprise-lb.yaml --days 120

from                to                  duration success rate
2024-11-19 02:30:46 2024-11-20 02:30:08 16m19s   🤨 75% 3/4
2024-11-17 18:06:02 2024-11-20 02:30:08 16m2s    🤨 50% 4/8
2024-11-15 02:30:38 2024-11-20 02:30:08 16m15s   🤨 69% 11/16
2024-11-09 18:05:42 2024-11-20 02:30:08 16m21s   🤨 72% 23/32
2024-10-30 02:28:35 2024-11-20 02:30:08 16m8s    🤨 77% 49/64

failed jobs
job name                                       failure count
Setup & Test (multi-node, 6.1-20240710.064909) 15
```

The big difference between the single-node and multi-node test is the
multi-node test doesn't cleanup containers or k8s resoures after tests.

Based on the observation that random tests are failing and only
multi-node tests are flaky, I came up with the assumption that the
containers or k8s resources left after the test consumes the GH runner's
resource and makes tests flaky.

To validate this assumption, we'll remove the -cleanup=false flag from
the multi-node tests as well. This loses the ability to collect sysdump
after the "Setup & Test" step, but in the previous PR, we implemented
the feature to collect sysdump within the test, so this shouldn't be a
problem.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>