bgpv2: Introduce ConflictingClusterConfig condition
Introduce a new status condition ConflictingClusterConfig to the
IsovalentBGPClusterConfig which indicates that the ClusterConfig is
selecting the node that is already selected by other ClusterConfig.

We reused an existing conflict detection mechanism in the Cilium
Operator which checks the existing NodeConfig's owner reference when it
tries to create the new NodeConfig. If there's another ClusterConfig
owning the same node, it is observed as a conflict.

Because of this detection mechanism, the condition in the single
ClusterConfig doesn't have the global view of the conflict relationship.
To get the complete view, users need to gather the observed conflicts
from multiple ClusterConfigs.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>