helm: Enable export aggregation by default when timescape is enabled
We enable export aggregation by default when using integrated Timescape.
This should significantly reduce the number of exported flows, so that
integrated Timescape runs well by default on larger clusters.

We enable aggregation per connection, as this should already
significantly reduce the number of exported flows, without losing a lot
of details.

We don't enable aggregation by default if the export-path is set by the
user. In that case we need to assume that the user already exports flows
to different locations, and we don't want to change the exiting export
behavior when enabling Timescape.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>