helm: add lower TTL for integrated Timescape
We lower the default TTL for integrated Timescape to be in line with
Timescape Lite. This should ensure that integrated Timescape runs well
by default on larger clusters.

We lower the default TTL even lower, if no persistant volume is
configured, and limit the emptyDir size to 5Gi to protect the host.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>