feat: Add reconciliation logic for updates and out-of-band changes
The controller reconciles now when updates have been applied
to the CiliumConfig, to the helm charts (by a new release of the
operator) or to the resources managed by the operator (out-of-band
changes). It makes sure that the state of Cilium in the cluster
reflects the intent captured in the custom resource for the version
 of the helm charts in use.
The logic uses now the helm libraries only to generate the manifests,
which are applied thanks to the controller-runtime client.
This allows to react to changes without hot looping.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>