enterprise/fqdn-proxy: fix release process
Adjust the release instructions to update the fqdn-proxy Helm chart for
v1.14-ce and earlier to match the ones for later versions. The incorrect
instructions lead to the image tags not being updated for the last few
1.14 CEE releases.

Also let Makefile targets always update chart and values by not letting
the target depend on files being changed. This makes sure to always
completely generate the charts and values when `make` is invoked, so we
don't accidentally miss updating one or the other on release.

These aren't expensive or long-running operations so it doesn't matter
to run them every time.

For isovalent/customer-support#1248

Signed-off-by: Tobias Klauser <tobias@isovalent.com>