enterprise/fqdn-proxy: update chart README.md when updating Chart/values
David reports that in addition to the image versions being outdated in
the Helm charts and values for some releases, also the README.md wasn't
regenerated to reflect the updated Helm chart. Make sure to keep the
README.md up-to-date by running the not-so-aptly-named test.sh script
which will take care of this.

Reported-by: David Bimmler <david.bimmler@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>