lb: fix udp health checks algorithm
The active health checker for UDP expects an echo of a datagram it
sends (namely, a character 't'). It considers only one type of error:
ICMP_PORT_UNREACHABLE (which is returned as ECONNREFUSED because the
socket is connected). However, it also considers a timeout as an
error.

Change two things in this algorithm: first, send an empty datagram.
This is a bit more reliable than sending a non-empty, as the
receiving application may not reply with an echo for a random
datagram (on the app port!). Seconly, for the same reason, do not
mark the timeout case as a failure.

(The right approach here is to switch to L7 healthcheks for UDP.)

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>