fix: move from helm upgrade action to install action
helm upgrade action even in dry run requires the helm secret
to be present if not configured in client only mode.
Client only mode is not possible due to lookup functions used
in helm charts.
Helm labels and annotations are required for resources adoption
and are added in the post renderer function.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>