gha: run SEGW test for PRs and on schedule
[ v1.15-ce commit e5ba2b98ebc243782da62a69751c410afba7cca3 ]

The blamed commit introduced a new E2E test covering standalone egress
gateway. However, it did not include it as part of the tests run for
each PR via the global ariane trigger, which is also used to identify
the tests to be run on schedule. Let's get that fixed.

Fixes: 88462981541d ("gha: standalone egress gateway test")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>