lb: add E2E test for T1-only udpProxy
This test is similar to a one for tcpProxy. It utilizes the new
v0.0.9 version of quay.io/isovalent-dev/lb-healthcheck-app adds a UDP
listener. The test is trying to reach the backend a few times, and
expects an echo, which `nc` outputs to a file. The test re-tries
a few times, since neither UDP nor nc are reliable.

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>