gha/panda: match export aggregation and hubble metrics settings
Ensure that these settings match the ones actually configured by the
customer, to provide consistent results. Specifically, they are
relevant in the context of CPU/memory usage, which is currently
significantly higher compared to that observed in sysdumps due to
the lack of filtering/aggregation before exporting flows (which
is a notoriously expensive operation). The hubble metrics are not
expected to introduce significant differences in this test, but
let's still enable them to better mimic the target environment.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>