gha/panda: use cert-manager to generate tls-interception certs
Leverage cert-manager and trust-manager to declaratively configure the
secrets used for TLS interception, for improved clarity and to guarantee
that the default CA bundle is actually up-to-date.

The Cilium full-tls-context=false setting is additionally configured
to prevent propagating the CA certificate part of the terminating TLS
secret to envoy, as that would enable client authentication. See [1]
for additional details.

[1]: 113673394f1b ("xds: Only pass relevant TLS config for originating/terminating TLS")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>