k8s,operator/bfd: Allow unnumbered BFD peers in BFD CRDs
To allow for unnumbered BFD peers aligned with BGP,
make PeerAddress optional in IsovalentBFDNodeConfig.
When PeerAddress is not specified, Interface must be specified,
in which case an auto-detected link-local
IPv6 neighbor address will be used for peering.

In IsovalentBFDNodeConfigOverride, do not use PeerAddress
to map override peer to NodeConfig peeer, but use
newly introduced Name field instead. PeerAddress field
is removed, as it has no use in the override now.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>