egwha: close sockets for CNI connections on unavailable gateway IPs.
Adds a new socket-termination feature to egress-gateway-ha.
When enabled, this forces client-connections being tracked by EGWHA
(i.e. connections being forwarded to an Egress Gateway to be masqueraded)
to be closed when the pinned egress gateway IP is no longer available.

This occurs when a EGWHA-CT map value's gatewayIP value is no longer
part of any matching policies healthygatewayips - indicating that
the connection was being tracked and forwarded to a gateway that
is no longer available (i.e. due to operator health checks on the node).

Connections are always terminated by destroying the socket on the
client side via netlink.

This solves the problem of stale connections, forwarded through
unavailable gateways, spending extended periods of time waiting for
packets before finally being timed out by the kernel.

Instead, a node healthcheck for selected egress-gateway-ha GW nodes
happens by the operator control plane every couple seconds. When
the operator determines that a Node is unavailable it is removed
from the healthyGatewayIPs list in the IsovalentEgressGatewayPolicy
status.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>