enterprise: add general purpose sockets package.
Introduces a enterprise `sockets` library that provides the
following families of functions:

note: the 'As' variant accepts a network namespace handle.

* Destroy{As,}: Destroy a socket.
* Iterate{As,}: Iterate sockets.

This code partially derives from OSS code in
`pkg/datapath/sockets/` as is mentioned in the license header comment.

Eventually, code in `pkg/datapath/sockets` can be rewritten on top
of the more general code in the enteprise sockets package and we
can contribute this back to OSS.

However, due to a lack of test coverage, it will be safer to
commit this as a seperate piece of code for use by egwha socket
termination.

As a follow up, we will want to stabilize the usage of the oss
library and provide more test coverage in service/connectivity
to ensure that refactoring that code using the more general Iterate{As,}
functions doesn't break the service LB connection termination
functionality.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>