gha/panda: don't delete test namespaces
Let's save a bit of time, considering that we are subsequently deleting
the cluster in any case. Most importantly, this ensures that the final
sysdump is captured with all the test pods up and running, rather than
potentially in an intermediate state following the deletion (e.g., some
of the identities may already having been GCed, while others not).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>