ilb/cli: Fix reporting of BGP Peers in LB status
Previously, the reporting was assuming that each BGP peer expects to
receive all BGP advertised routes. The assumption is wrong, because with
the BGPv2 we can control which peers can receive what advertisements.

This led to invalid "BGP Peers" reporting. For example, after running
some of the ILB E2E tests - "DEG  [2/106]" when all peers/BGP
connections were healthy.

Fix this by figuring out which routes and peers are relevant for a given
service from IsovalentBGPClusterConfigs. A more proper fix would be to
backtrace the peers of the service from IsovalentBGPAdvertisements =>
IsovalentBGPPeerConfig => IsovalentBGPClusterConfigs.

Signed-off-by: Martynas Pumputis <m@lambda.lt>