lb: grafana: t2.json: sync with lb.json changes
At first there were T1 and T2 dashboards, then some of the T2 panels got
incorporated in the T1 dashboard (which got renamed to just ILB).

As some of the ILB panels got updated, bring their version in the T2
dashboard up to date with their counterpart in the ILB dashboard.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>