ci: Add a workflow to build lifecycle operator
The lifecycle operator gets rebuilt each time a PR or
a push is made to the /olm directory. This covers cases
where the directory containing the helm charts for the
lifecycle operator are synced and code changes made
directly to the operator logic.
Multi-arch builds and SBOM generation are not implemented yet.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>