gha: always reset values in fqdn-ha upgrade/downgrade workflow
439f5584cbb9 ("CLI: cilium upgrade preserver previous configurations")
changed the default behavior of `cilium upgrade` to preserve the
previous configuration upon upgrade. While this is the expected behavior
in most situations, it turned out being problematic in the fqdn-ha upgrade
downgrade tests. Indeed, the images are not overwritten for the stable
configuration, leading in turn to not actually performing the downgrade.
Let's fix this by explicitly resetting the values when performing the
upgrade and downgrade operations.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>