gha/panda: configure beefier control-plane nodes
In preparation for the subsequent increase of the worker nodes count
to 1K, let's configure three control plane node, doubling their specs
as well. Similarly, we double the specs of the worker node used to
host Prometheus, and disable the synchronization of cloud routes to
reduce the amount of API calls, given that we deploy Cilium in
tunneling mode. Finally, let's increase the size of the etcd disks
(and ensure they are of type ssd), to improve the attainable IOPS,
and double the maximum number of inflight requests.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>