gha/panda: simplify wait for workload nodes readiness
Currently, we wait for workload nodes readiness leveraging the dedicated
kops command, which in the past proved to be expensive in terms of API
calls. In preparation for the subsequent increase of the worker nodes,
let's replace it with a bespoke logic that directly checks Kubernetes
nodes readiness. While being there, let's also uniform the wait logic
of the etcd instance group (we don't validate readiness there because
Cilium has not been installed yet).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>