gha/panda: configure nodes filter for sysdump capture
Currently, we always include the information of Cilium agents running on
all nodes when capturing a sysdump. In preparation for increasing the
number of worker nodes, let's configure a node filter, so that we
ensure that the resulting sysdump is manageable. Specifically, the
intermediate sysdump includes information from all control-plane nodes
and 10 workload nodes, while the finaly sysdump additionally includes
information for nodes hosting an agent which crashed during the test.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>