hubble-timescape: Reduce clickhouse system table disk usage
By default, most ClickHouse system log tables don't have a TTL at all,
and the once that do, have a TTL of multiple months. This means, system
log tables can eventually take up a significant amount of disk space.
This is especially relevant when the flows table TTL is low, which is
the case for integrated Timescape.

We do two things to limit the sytem table disk usage:

* Disable some system log tables that we don't expect to use
* Keep the log tables that are useful for debuging, but set a low TTL of
  30min.
* Set a longer TTL for error and crash logs. These are invaluable when
  debugging customer issues and they usually stay very small

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>