ilb: Fix health-check restoration bug
When testing I noticed that upon a quarantined backend the agent is being
restarted, and the health-checker resumes operation, it never moves the
backend out of quarantined state despite the health check declaring the
probe packet a success.

The problem is that UpsertService() missed to insert the backend into the
unhealthyBackends set, and therefore the health checker's internal state
was set to active, and given the health probes were a success, the
health.probe.healthy != newProbe.healthy test in sendHealthProbes was
always false.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>