ilb: add connectionfiltering to UDPRoute
This commit adds the possibility to configure
connection-filtering based on the source ip on
the UDProute

Note: It's intentional that we introduce quite a bit
of boilerplate with the TCP proxy pendent. But it's
easier to handle model changes going forward, if the
two features evolve differently (without forgetting
to implement it for both features).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>