gha/panda: don't label etcd nodes as cilium.io/no-schedule
By default, this label does not prevent Cilium agents from being
scheduled, while it prevents the scheduling of cilium-envoy pods.
While we could configure node affinities to let all Cilium-related
components respect it, that would cause the subsequent cluster
readiness check to fail as the etcd nodes are not marked as ready.
Considering that it doesn't actually matter for this test, let's
just play it easy and let all Cilium pods run on etcd nodes as well.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>