ilb: update health check app and use new UDP server response
This commit updates the health check app to `v0.10.0` that changes
the response of the udp server to include the service- and instancename
in addition to echoing the request.

-> `<service-name>:<instance-name>:<echo>`

This is done in preparation to check for UDP sessions (to be able
to verify that a UDP packets of the same "session" (4 tuple) are
always sent to the same backend).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>