cilium-cli: Correlate scenarios with code owners
[ upstream commit a87f5b6e93d4758963b077e0e55bd318d6d101da ]

Correlate the test scenarios with the GitHub teams that own these test
scenarios, so that we can pretty-print some helpful assistance to
developers in CI when a test fails.

The intent here is that this makes the developer experience easier by
connecting developers who are struggling with a CI failure with the
owners of those tests.

Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Joe Stringer <joe@isovalent.com>