ilb: remove udp proxyaccesslog on udp proxy
Currently, the fields `accessLog` & `proxyAccessLog` are configured on the
Envoy UDPProxy.

According to the Envoy docs, it seems that `accessLog` is for UDP session
based logging while `proxyAccessLog` for the proxy in general.

In reality it seems that `proxyAccessLog` isn't logging anything until the
configuration is destructed.

Therefore, this commit removes the UDP `proxyAccessLog` on the UDPProxy configuration.

The `accessLog` field is still set - which is the accesslog that configures
the UDP session access log.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>