gha/panda: add selector to test services
Currently, the test services created as part of the Panda scale test
are not associated with a selector, given that the endpointslices are
created manually. However, Cilium potentially treats services without
a selector differently; while this allowed up to discover an existing
inefficiency in policy rule translation, it may hide other problems
that are actually specific to services with a selector, and that Panda
would hit as well. Hence, let's add a selector to the service template,
and configure the endpointslices to be owned by a different (non
existing) controller, so that they don't get reconciled by the
controller manager.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>