enterprise/cilium-cli: add Dockerfile
This will be used to build enterprise cilium-cli container image
releases. Note that contrary to the OSS Dockerfile in
cilium-cli/Dockerfile this one doesn't install the cloud CLIs which are
needed for cluster configuration auto-detection. The only supported
commands in enterprise release are `cilium sysdump` and `cilium
connectivity test`, both of which don't require that auto-detection and
thus the cloud CLIs can be omitted.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>