fqdn-proxy/api: Add Agent Proxy Status Stream
"fqdn-proxy" needs a fuller view into the status and
configuration of the DNSProxy running on the agent to
utilize more efficient and resilitent methods of
implementing the DNSProxy itself, such as reading directly from
the IP cache bpf map (rather than using a gRPC call) and knowing
when to reload the bpf ipcache on agent restart.

This ProxyStatus will grow larger in subsequent commits,
including reporting on endpoint regeneration status,
but for now it reports the agent proxy uptime and if
it supports "V1" bpf IPCache keys.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>