gha/panda: move node CIDR to the bottom of the range
The default node CIDR (10.16.0.0/16) can conflict with the pod CIDRs
assigned by the cluster-pool IPAM depending on the number of nodes
--- ~4K --- with the current settings and assuming that they are
drawn sequentially. Hence, let's move the node CIDR to the bottom
of the range, to prevent funky problems. Let's also explicitly
ensure that the IPAM range is non-overlapping, as an extra safety
measure in case it were ever changed to not pick CIDRs in order.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>