fqdn-proxy: Add Flag to Enable Offline Mode
Add a flag to enable offline mode. This is needed so that
offline moded can be disabled. In certain deployement
contexts, such as Openshift, cross mounting the bpf
file system into an unprivileged Daemonset is not
allowed by SELinux. The performance and resiliency
features that rely on accessing the bpf filesystem
need to be disabled in these situations.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>