ipmigration: Add `enterprise/pkg/ipmigration`
This is the first of a series of commits adding a new `ipmigration`
package to the enterprise codebase. The ipmigration package is
responsible for supporting IP migration by allowing pods to be attached
and detached from Cilium. A detached pod is a pod is a pod without a
corresponding Cilium endpoint (and without connectivity), but allows the
IP to be used concurrently by a different pod. This is useful in the
context of KubeVirt VM migration where the target VM can be set up in a
detached state, and become attached once migration has finished the
source VM pod has been deleted, thereby also allowing ownership of the
source VM pod IP to be transferred to the target pod.

Pods are marked as detached by an annotation. This commit contains the
name and value for the annotation. Later commits will add the actual
logic to start pods in a detached state and moving pods from a detached
state to an attached state.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>