ipmigration: Intercept API calls invoked for CNI ADD
This commit adds the first half of the main business logic of the
`ipmigration` package, namely the part which intercepts the Cilium API
calls invoked by cilium-cni during CNI ADD: `POST /ipam` and `PUT
/endpoint/{id}`.

When they observe IPAM allocation or endpoint creation requests for a
detached pod, they will return a fake response without actually creating
an endpoint. This allows the pod to become ready (but without network
connectivity).

In a second commit, we will then add the logic which allows these
detached pods to become attached.

If the IP migration feature is enabled via an agent flag, the API
handlers are injected via `cell.DecorateAll`. This allows us to
overwrite the behavior of the upstream handlers. Requests for
non-detached pods are forwarded without modifications.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>